refactor: move step-specific creation components into their CreationSteps folders
