+++ src/components/display/GameServer/CreateGameServer/CreationSteps/Step2/TemplateVariableForm/index.ts
