+++ src/components/display/GameServer/CreateGameServer/CreationSteps/Step2/TemplateVariableForm/inputs/index.ts
